requirements.txt
google-cloud-storage
numpy
pandas
scikit-learn
tensorflow
transformers